+++ src/main/java/com/shootdoori/match/dto/TeamReviewResponseDto.java
@@ -3,13 +3,13 @@ package com.shootdoori.match.dto;
 import com.shootdoori.match.entity.TeamReview;
 
 public record TeamReviewResponseDto(Long teamReviewId,
-                                    Integer matchId,
+                                    Long matchId,
                                     Long reviewerTeamId,
                                     Long reviewedTeamId,
-                                    Integer rating,
+                                    Long rating,
                                     String comment,
-                                    Integer punctualityRating,
-                                    Integer sportsmanshipRating) {
+                                    Long punctualityRating,
+                                    Long sportsmanshipRating) {
 
     public static TeamReviewResponseDto from(TeamReview review) {
         return new TeamReviewResponseDto(
